requirements.txt
-e git+https://github.com/Sprtch/despinassy.git#egg=despinassy-pkg-tperale
psycopg2
pytest
python-daemon
evdev==1.3.0
redis
pyserial
PyYAML
